backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart
pydantic>=2.0.0
pydantic-settings

# Claude API
anthropic

# Database
sqlalchemy>=2.0.0
aiosqlite

# Utilities
python-dotenv
gitpython

# Async support
aiofiles
httpx

# Logging
structlog
